Kernel: Allow higher audio sample rates than 65kHZ (`u16`)
Executing `asctl set r 96000` no longer results in weird sample rates
being set on the audio devices. SB16 checks for a sample rate between 1
and 44100 Hz, while AC97 implements double-rate support which allows
sample rates between 8kHz and 96kHZ.

+++ Userland/Services/AudioServer/ClientConnection.cpp
@@ -73,7 +73,7 @@ Messages::AudioServer::GetSampleRateResponse ClientConnection::get_sample_rate()
     return { m_mixer.audiodevice_get_sample_rate() };
 }
 
-void ClientConnection::set_sample_rate(u16 sample_rate)
+void ClientConnection::set_sample_rate(u32 sample_rate)
 {
     m_mixer.audiodevice_set_sample_rate(sample_rate);
 }
